Add AtomInfo.get_building_block_atom() (#267)
This commit introduces two new public methods:

    * ``AtomInfo.get_building_block_atom()``
    * ``Molecule.get_canonical_atom_ids()``

``AtomInfo.get_building_block_atom()`` is introduced to solve issue
#260. Specifically, this commit makes it possible for users to identify
which ``Atom`` of a ``ConstructedMolecule`` originates from which
``Atom`` of a ``BuildingBlock``, by calling
``AtomInfo.get_building_block_atom()``. Adding
``Molecule.get_canonical_atom_ids()`` is a necessary evil for the
correct implementation of this feature, at least as far as I can tell.
Here is why:

    * ``ConstructedMolecule.with_canonical_atom_ordering()`` must work.
    * When ``ConstructedMolecule.with_canonical_atom_ordering()`` is
      called, it also calls ``Molecule.with_canonical_atom_ordering()``
      on its building blocks. This means the building blocks of a
      canonical constructed molecule are replaced with canonical
      building blocks.
    * After ``ConstructedMolecule.with_canonical_atom_ordering()`` is
      called, ``AtomInfo.get_building_block()`` points to the new,
      canonical building block.
    * Similarly ``AtomInfo.get_building_block_atom()`` needs to point
      to the atom in the new, canonical building block.
    * The implementation of
      ``ConstructedMolecule.with_canonical_atom_ordering()`` which
      creates new ``AtomInfo`` instances for the new, canonical
      constructed molecule, needs to take the ``Atom`` from the old
      ``AtomInfo`` and figure out what the corresponding ``Atom`` in
      the new, canonical building block is. For this, the building
      block needs to provide a map, which maps the old atom id to the
      canonical atom id.
    * In theory, ``Molecule.get_canonical_atom_ids()`` could have been
      added as a private method because ``ConstructedMolecule``
      inherits ``Molecule``. However, I think this is a bad idea.
      The key point is that ``Molecule`` is an abstract base class, and
      private methods are explicitly excluded from its interface. They
      therefore cannot be assumed to exist on ``Molecule``  objects.
      Therefore, the implementation of
      ``ConstructedMolecule.with_canonical_atom_ordering()``
      cannot call ``Molecule._get_canonical_atom_ids()`` on its
      building blocks, because a private method is not part of the
      interface of ``Molecule``, and therefore is not guaranteed to
      exist.

In order to implement ``AtomInfo.get_building_block_atom()`` the
serialization and deserialization of ``AtomInfo`` also needed to be
updated. The serialized representation of an ``AtomInfo`` instance
needs to preserve the building block atom. Currently, only the JSON
serialized  form exists, so only this needed to be updated. When an
``AtomInfo``  instance is serialized into JSON, it is represented as an
array. The building block atom held by an ``AtomInfo`` instance was
added to the end of this array in order to maintain backwards
compatibility. By adding the building block atom at the end of the
array, the other data in the array can still be accessed at the same
index as before.

I will note that there is a slight inefficiency in the
implementation of
``ConstructedMolecule.with_canonical_atom_ordering()``
this commit introduces, which I'm not
too happy about. Canonical ordering is computed twice for each building
block. Once when ``with_canonical_atom_ordering()`` is called on it,
and once when ``get_canonical_atom_ids()`` is called. This can be
addressed in two ways as far as I can see:

    * Cache the ordering on first computation. Costs space.
    * Change the interface, such that when
      ``with_canonical_atom_ordering()`` is called, it returns both the
      canonical clone, as well as the map from old to new id.

I prefer option two, because there is no performance penalty in time or
space. Potentially, this could be added as a separate method, so we
would end up with 1 method for returning just clone, 1 for returning
just map of ids, 1 for returning both together. I've added issue #270
to discuss this further.

Naturally, this commit raises the question about the existence of
``BondInfo.get_building_block_bond()``. I don't think
``BondInfo.get_building_block_bond()`` is required to solve issue #260,
and adding it is therefore out of scope for this commit. However I do
think this method should be  added in the future. Having
a mapping to the original building block bond is potentially useful to
users and there is no good reason that it needs to be lost. I added
issue #271 to log this.

The main source code changes are as follows:

    * ``atom_info.py``: Add
      ``AtomInfo.get_building_block_atom()``.
    * ``molecular/.../constructed_molecule.py``: Update
      ``ConstructedMolecule.with_canonical_atom_ordering()`` so that
      when a canonical constructed molecule is created, it contains
      ``AtomInfo`` instances which point to the atoms in the canonical
      building block.
    * ``molecule.py``: Add
      ``Molecule.get_canonical_atom_ids()``, because it is necessary
      to implement
      ``ConstructedMolecule.with_canonical_atom_ordering()``.
    * ``deletions_summary.py``: Add the building block atom when
      constructing ``AtomInfo`` instances.
    * ``placements_summary/atom_batch.py``: Add the building block atom
      when constructing ``AtomInfo`` instances.
    * ``reactions_summary/atom_batch.py``: Add the building block atom
      when constructing ``AtomInfo`` instances.
    * ``json/deserializers/utilities.py``: Make sure you
      extract the building block atom from index 2 of a JSON array
      which represents an ``AtomInfo`` instance. Add the building
      block atom when constructing ``AtomInfo`` instances.
    * ``json/serializers/.../constructed_molecule.py``: Add the
      building block atom id to the end of the array which forms the
      JSON representation of an ``AtomInfo`` instance.