Update rdkit (#325)
Related Issues: #309 

Updates the `rdkit` installation instructions to use `conda-forge`
instead of the `rdkit` channel. Also updates the test 
environment to use the latest version of  `rdkit` and other 
dependencies. Finally, the position matrices of test molecules have 
been updated because of numerical changes resulting from the
version updates. The structures of the molecules were checked 
manually and seem OK.