Property vector caching (#326)
Related Issues: #321 

The `PropertyVector` fitness calculator had `input_database` and
`output_database` arguments in its initializer, as well as examples
saying that it allows the caching of fitness values. However, the 
implementation of `get_fitness_value()` did not do any caching.

In order to prevent issues like this in the future I've added a new
test for fitness calculators, which tests that fitness values do get
cached.