Plug leak (#339)
Related Issues: #307 

When `PeriodicInfo` is being initialized it is not making a copy of
the arrays it is being given. This means that if those arrays get
changed on the outside, the internal state of `PeriodicInfo` also gets
corrupted. Similarly, when the `get` methods are called, the internal
state is returned directly, rather than a copy of the arrays. If
someone calls `get` and then changes the value in the returned array,
they will also change the internal state of `PeriodicInfo`.

This commit just makes sure to make a copy of the arrays which are
received as input and returned as output.