Round svd results (#343)
The results of using `np.linalg.svd` will get rounded to 14 decimal
places to prevent errors associated with non-deterministic results from
this function.

The non-determinism of this function was discovered because in CI
the tests for the `M8L6Cube` cage would fail sometimes, even when
run in a Docker container, which should provide a completely consistent
environment. I went through the call stack during the construction of
the cage, to find at what point the numerical results were different
between a successful and failed run. I found that for Vertex with an
id of 8, the building block getting placed ended up having slightly
different positions sometimes. The cause of this was the svd function
in `Molecule.get_plane_normal()`, which would sometimes return a
slightly different value in one of the last decimal points. I assume
that this little difference ended up getting amplified through the
successive operations.

Based on this issue https://github.com/scipy/scipy/issues/7878, I
assume that this non-determinism is unavoidable.

To fix this I'm just going to round the results so that these small
differences get removed. I'm also updating all the position
matrices because I expect the rounding to change them slightly.