Allow multiple guests in host-guest complex (#352)
Related Issues: #351

Adds support for multiple guests to the `host_guest.Complex` topology
graph. This is implemented by adding a new parameter `guests` and
deprecating the `guest`, `start_vector`, `end_vector` and
`displacement` parameters. Instead, `guests` is an iterable of `Guest`
instances, each of which specifies the rotation and displacement for
that guest. The deprecated parameters will be optional for now, as will
the `guests` parameter, but a breaking change enforcing the switch will
happen in 6 months.

Adding a separate host guest complex topology graph, which allows
multiple guests, was decided against because it would likely add an
unnecessary maintenance burden.