Update CI
This commit does a couple of CI changes. First the building image
used by ReadTheDocs is set to "testing", so that Python 3.9 can be
used to compile the docs. This means that the compiled docs correctly
handle types like `tuple[int, int, int]`. This wasn't being handled
correctly even though the `from __future__ import annotations` import
was used.

Also updates the test docker image to include `vabene`, because
`vabene` is going to be used for type hinting and therefore needs to be
added as an explicit dependency.