Add Spinner (#356)
This commit adds a new optimizer, `Spinner`, which is specialized for
the optimization of host-guest pairs. This is most crucial for systems
with multiple guests that may have overlapping structures. The
implementation follows the same form as the `MCHammer` optimizers and
requires a new dependency `SpinDry` which carries out the actual
optimization.