Stop mocking modules
Because of the introduction of type hints, libraries can no longer be
mocked for the purposes of documentation. This is because the the types
in the library need to actually imported for the documentation
compilation to work. As a result, this commit removes all things
related to the mocking of modules from the repo. This includes the
requirements file `docs/requirements.txt` because all the dependencies
already exist in `tests/docker/environment.yml`. The dependencies are
there already because we use doctest, so any dependencies required
for Sphinx are already present in the test environment.

The `MOCK_MODULES` variable has been removed from
`.github/workflows/tests.yml` because it has been removed from
`docs/source/conf.py` because the mocking has been removed.