Add BuildingBlock.get_num_placers
Related Issues: #376

A new method, `BuildingBlock.get_num_placers` has been added. This
is necessary because sometimes the number of placer atoms is necessary,
for example, by unaliging vertices, which needs to check the number of
placer atoms, in order to determine if the unaliging vertex should be
used. Previously all the placer ids would have to be iterated though
with `BuildingBlock.get_placer_ids()`, which is both ugly and less
performant.