Add benchmarks (#389)
This commit adds a new folder for holding benchmarks, which can be
run with `pytest benchmarks`. The `pytest.ini` file has been updated
so that pytest is able to collect files and functions with the
`benchmark` prefix. This just means that benchmarks can be called
`benchmark_something` rather than having to call them `test_something`.

Currently I'm only adding a single benchmark for the kagome COF but
more can be added as the need arises.