Fix `get_all` for Mongo dbs (#368)
Related Issues: #366, #369

This commit ensures that `.get_all()` methods in MongoDB databases
will correctly return all molecules. Previously, the implementation
went through all the entries in the position matrix collection, and
tried to lookup the corresponding molecule in the molecules collection.
If the corresponding molecule was not found, a `KeyError` would be
raised. However, this would cause a problem in cases where the
had two molecule collections and a shared position matrix collection.

The current implementation will go through all the entries in the
molecule collection and do a lookup in the position matrix collection.
If there is no entry in the position matrix collection, the molecule
entry will simply be skipped. This skipping is done as part of the
MongoDB aggregation pipeline for improved performance.

Co-authored-by: Lukas Turcani <lukasturcani93@gmail.com>