Use set for placer ids. (#390)
Related Issues: #372 , #276, #156

Found, while working on #372, that `BuildingBlock._placer_ids`
contained duplicate ids from distinct functional groups. This PR fixes
this issue by using a frozen set in `building_block.py`.

Also added a test for this specific issue.

Also added the `BuildingBlock._core_ids` attribute and the
`BuildingBlock._get_core_ids()` method such that this property, similar
to `_placer_ids` is set on initialization of the `BuildingBlock`
instead of collected from scratch on call of
`BuildingBlock.get_core_atom_ids()`. The reason for this is that the
core atoms also need to be deduped, so its better to do it once,
during initialization.

Many position matrices change negligibly but enough to fail the tests
due to new centroid positions due to the set.

There is also an unrelated change of removing the initializer from the
Cage and Cof vertices. This is to unify the interface of the vertices.
It also means that a redunant fixture was removed: `unaliged_cof1`.