Use unaligning vertex in Linear polymer (#393)
Related Issues: #156

This commit allows the user to place molecules like `BrCBr` on the
`polymer.linear` topology graph using `UnaligningVertex`. This is
handled by the `Linear` initializer when creating vertices: all
vertices with building blocks that only have one placer atom are
converted to an `UnaligningVertex` - same as in COFs and cages.

Two new tests added for vertex and molecule fixture.