Add type hints (#398)
Adds type hints, which is self-explanatory. The more important thing is
that the `molecule` folder got moved from `molecule` into `molecular`.
The reason for this is to avoid a cyclic dependency which exists
because `building_block.py` depends on `FunctionalGroupFactory` while
`FunctionalGroupFactory` depends on `molecule.py` this means that
`molecule.py` cannot be in the same submodule (i.e. `molecules`) as
`building_block.py` and has therefore been moved directly into
`molecular`.