Add type hints (#401)
Add type hints for molecular writers.

I'm also abandoning the qualified imports because they can lead to
bad imports which mess up formatting in the compiled docs. For
example doing `from .. import molecule as _molecule`` ended up
importing the wrong module by accident. This failed silently but then
the compiled docs looked off. So its safer to just always import
`Molecule` because I think that will lead to an error if it is not
found.