Update test environment and Spindry version (#422)
Reviewers: @lukasturcani

This commit just updates the usage of `Spindry` in `stk` to fit the new
interface. Only a minor change was required, there is no longer the
ability to change the sigma values used in `stk.Spinner()` because
these are now set to the atom radii based on
`stk.Atom.__class__.__name__` (this was already handled in the
interface).