Update test environment (#437)
To keep `stk` update I'm updating all the dependencies in the test
environment. As a result of this, I've found that the
`pandas.DataFrame.append()` API is deprecated, so I've updated to the
new `pandas.concat()` API.

Also, because `rdkit` can now be installed from `pypi` I've dropped the
references to `conda install -c conda-forge rdkit`.

I've also removed the old `get_periodic_info()` API, which has had a
use warning for a long time.