Add vertex positions argument to cage topologies (#513)
This commit will add a `vertex_positions` argument to cage topology 
graphs. This allows the users to overwrite the scaling and positioning
of any vertex (using a similar `dict[int, position]` format as the
`vertex_alignments` arguments).

It will not work for `OnePlusOne` graph currently because of strict edge
positioning.

Indeed. some choices in the code changes stem from me trying to avoid
editing code outside of the cage namespace.

A few type hints were added too on any file I edited.

I have added tests for a few cage graphs, but fundamentally it is hard
to test the success or failure of this except for consistency, which I
believe is working.