Fix matplotlib backend (#542)
Related Issues: #541 

As mentioned in the issue, this commit removes the setting of the
matplotlib backend, which we presume is due to an out-dated decision for
the evolutionary algorithm.

The output of:

```python
>>> import matplotlib
>>> print(matplotlib.get_backend())

>>> import stk
>>> print(matplotlib.get_backend())
```
 
is now (on pop-os):
```
tkagg
tkagg
```


This commit also makes small changes associated with `ruff`, and
an rdkit version constraint to avoid issues with metal-containing smiles
strings.